renames entitylinkedlist to heropool
