+++ module/mempool/stdmap/backdata/heropool/entitiesLinkedList_test.go
@@ -1,4 +1,4 @@
-package arraylinkedlist
+package heropool
 
 import (
 	"fmt"
